Put all recording hook functions in a single .h
